events/requirements.txt
pusher
aiohttp
fastapi
uvicorn
python-multipart
python-dotenv
